Refactor & Lint (#525)
* move function calls

* fix lint
