+++ cmd/selectmany/selectmany.go
@@ -19,8 +19,10 @@ import (
 	sf "github.com/snowflakedb/gosnowflake"
 )
 
-var cpuprofile = flag.String("cpuprofile", "", "write cpu profile to file")
-var memprofile = flag.String("memprofile", "", "write memory profile to this file")
+var (
+	cpuprofile = flag.String("cpuprofile", "", "write cpu profile to file")
+	memprofile = flag.String("memprofile", "", "write memory profile to this file")
+)
 
 // getDSN constructs a DSN based on the test connection parameters
 func getDSN() (string, *sf.Config, error) {
@@ -95,18 +97,11 @@ func run(dsn string) {
 		log.Fatalf("failed to run a query. %v, err: %v", query, err)
 	}
 	defer rows.Close()
-	var v1 int
-	var v2 int
-	var v3 int
-	var v4 int
-	var v5 int
-	var v6 int
-	var v7 int
+	var v1, v2, v3, v4, v5, v6, v7 int
 	fmt.Printf("Fetching the results. It may take long. You may stop by Ctrl+C.\n")
 	counter := 0
 	for rows.Next() {
-		err := rows.Scan(&v1, &v2, &v3, &v4, &v5, &v6, &v7)
-		if err != nil {
+		if err = rows.Scan(&v1, &v2, &v3, &v4, &v5, &v6, &v7); err != nil {
 			log.Fatalf("failed to get result. err: %v", err)
 		}
 		if counter%10000 == 0 {
