+++ restful.go
@@ -43,11 +43,10 @@ const (
 	heartBeatPath            = "/session/heartbeat"
 )
 
-// FuncGetType httpclient GET method to return http.Response
-type FuncGetType func(context.Context, *snowflakeRestful, *url.URL, map[string]string, time.Duration) (*http.Response, error)
-
-// FuncPostType httpclient POST method to return http.Response
-type FuncPostType func(context.Context, *snowflakeRestful, *url.URL, map[string]string, []byte, time.Duration, bool) (*http.Response, error)
+type (
+	funcGetType  func(context.Context, *snowflakeRestful, *url.URL, map[string]string, time.Duration) (*http.Response, error)
+	funcPostType func(context.Context, *snowflakeRestful, *url.URL, map[string]string, []byte, time.Duration, bool) (*http.Response, error)
+)
 
 type snowflakeRestful struct {
 	Host           string
@@ -64,8 +63,8 @@ type snowflakeRestful struct {
 
 	FuncPostQuery       func(context.Context, *snowflakeRestful, *url.Values, map[string]string, []byte, time.Duration, uuid.UUID, *Config) (*execResponse, error)
 	FuncPostQueryHelper func(context.Context, *snowflakeRestful, *url.Values, map[string]string, []byte, time.Duration, uuid.UUID, *Config) (*execResponse, error)
-	FuncPost            FuncPostType
-	FuncGet             FuncGetType
+	FuncPost            funcPostType
+	FuncGet             funcGetType
 	FuncRenewSession    func(context.Context, *snowflakeRestful, time.Duration) error
 	FuncPostAuth        func(context.Context, *snowflakeRestful, *url.Values, map[string]string, []byte, time.Duration) (*authResponse, error)
 	FuncCloseSession    func(context.Context, *snowflakeRestful, time.Duration) error
@@ -182,8 +181,7 @@ func postRestfulQuery(
 		return data, err
 	}
 
-	err = sr.FuncCancelQuery(context.TODO(), sr, requestID, timeout)
-	if err != nil {
+	if err = sr.FuncCancelQuery(context.TODO(), sr, requestID, timeout); err != nil {
 		return nil, err
 	}
 	return nil, ctx.Err()
@@ -211,7 +209,6 @@ func postRestfulQueryHelper(
 	var resp *http.Response
 	fullURL := sr.getFullURL(queryRequestPath, params)
 	resp, err = sr.FuncPost(ctx, sr, fullURL, headers, body, timeout, true)
-
 	if err != nil {
 		return nil, err
 	}
@@ -220,14 +217,12 @@ func postRestfulQueryHelper(
 	if resp.StatusCode == http.StatusOK {
 		logger.WithContext(ctx).Infof("postQuery: resp: %v", resp)
 		var respd execResponse
-		err = json.NewDecoder(resp.Body).Decode(&respd)
-		if err != nil {
+		if err = json.NewDecoder(resp.Body).Decode(&respd); err != nil {
 			logger.WithContext(ctx).Errorf("failed to decode JSON. err: %v", err)
 			return nil, err
 		}
 		if respd.Code == sessionExpiredCode {
-			err = sr.renewExpiredSessionToken(ctx, timeout, token)
-			if err != nil {
+			if err = sr.renewExpiredSessionToken(ctx, timeout, token); err != nil {
 				return nil, err
 			}
 			return sr.FuncPostQuery(ctx, sr, params, headers, body, timeout, requestID, cfg)
@@ -239,44 +234,21 @@ func postRestfulQueryHelper(
 			ctx = WithQueryIDChan(ctx, nil)
 		}
 
-		var resultURL string
 		isSessionRenewed := false
-		noResult := isAsyncMode(ctx)
 
 		// if asynchronous query in progress, kick off retrieval but return object
-		if respd.Code == queryInProgressAsyncCode && noResult {
-			// placeholder object to return to user while retrieving results
-			rows := new(snowflakeRows)
-			res := new(snowflakeResult)
-			switch resType := getResultType(ctx); resType {
-			case execResultType:
-				res.queryID = respd.Data.QueryID
-				res.status = QueryStatusInProgress
-				res.errChannel = make(chan error)
-				respd.Data.AsyncResult = res
-			case queryResultType:
-				rows.queryID = respd.Data.QueryID
-				rows.status = QueryStatusInProgress
-				rows.errChannel = make(chan error)
-				respd.Data.AsyncRows = rows
-			default:
-				return &respd, nil
-			}
-
-			// spawn goroutine to retrieve asynchronous results
-			go getAsync(ctx, sr, headers, sr.getFullURL(respd.Data.GetResultURL, nil), timeout, res, rows, cfg)
-			return &respd, nil
+		if respd.Code == queryInProgressAsyncCode && isAsyncMode(ctx) {
+			return sr.processAsync(ctx, &respd, headers, timeout, cfg)
 		}
 		for isSessionRenewed || respd.Code == queryInProgressCode ||
 			respd.Code == queryInProgressAsyncCode {
 			if !isSessionRenewed {
-				resultURL = respd.Data.GetResultURL
+				fullURL = sr.getFullURL(respd.Data.GetResultURL, nil)
 			}
 
 			logger.Info("ping pong")
-			token, _, _ := sr.TokenAccessor.GetTokens()
+			token, _, _ = sr.TokenAccessor.GetTokens()
 			headers[headerAuthorizationKey] = fmt.Sprintf(headerSnowflakeToken, token)
-			fullURL := sr.getFullURL(resultURL, nil)
 
 			resp, err = sr.FuncGet(ctx, sr, fullURL, headers, timeout)
 			if err != nil {
@@ -291,8 +263,7 @@ func postRestfulQueryHelper(
 				return nil, err
 			}
 			if respd.Code == sessionExpiredCode {
-				err = sr.renewExpiredSessionToken(ctx, timeout, token)
-				if err != nil {
+				if err = sr.renewExpiredSessionToken(ctx, timeout, token); err != nil {
 					return nil, err
 				}
 				isSessionRenewed = true
@@ -336,8 +307,7 @@ func closeSession(ctx context.Context, sr *snowflakeRestful, timeout time.Durati
 	defer resp.Body.Close()
 	if resp.StatusCode == http.StatusOK {
 		var respd renewSessionResponse
-		err = json.NewDecoder(resp.Body).Decode(&respd)
-		if err != nil {
+		if err = json.NewDecoder(resp.Body).Decode(&respd); err != nil {
 			logger.WithContext(ctx).Errorf("failed to decode JSON. err: %v", err)
 			return err
 		}
@@ -468,8 +438,7 @@ func cancelQuery(ctx context.Context, sr *snowflakeRestful, requestID uuid.UUID,
 	defer resp.Body.Close()
 	if resp.StatusCode == http.StatusOK {
 		var respd cancelQueryResponse
-		err = json.NewDecoder(resp.Body).Decode(&respd)
-		if err != nil {
+		if err = json.NewDecoder(resp.Body).Decode(&respd); err != nil {
 			logger.WithContext(ctx).Errorf("failed to decode JSON. err: %v", err)
 			return err
 		}
