+++ result.go
@@ -9,8 +9,6 @@ import (
 type queryStatus string
 
 const (
-	// QueryStatusWaiting denotes a query execution waiting to happen
-	QueryStatusWaiting queryStatus = "queryStatusWaiting"
 	// QueryStatusInProgress denotes a query execution in progress
 	QueryStatusInProgress queryStatus = "queryStatusInProgress"
 	// QueryStatusComplete denotes a completed query execution
