+++ s3_storage_client_test.go
